+++ CHANGELOG.md
@@ -3,6 +3,8 @@ SPDX-FileCopyrightText: Copyright © DUNE Project contributors, see file LICENSE
 SPDX-License-Identifier: LicenseRef-GPL-2.0-only-with-DUNE-exception
 -->
 
+- Provide `communicate()` method in `OneDGrid` and `IdentityGrid`
+
 # Master (will become release 2.9)
 
 - UGGrid is now thread safe on the grid view.